refactor(row-adapters): split row adapters by domain
