+++ docs/architecture.md
@@ -463,7 +463,7 @@ Beyond the client-owned runtime graph, several feature APIs are implemented via 
 | `ArtifactDownloadService` | [`_artifact_downloads.py`](../src/notebooklm/_artifact_downloads.py) | Asynchronous download coordinator for finished artifacts. |
 | `_artifact_formatters` | [`_artifact_formatters.py`](../src/notebooklm/_artifact_formatters.py) | Markdown, HTML, and plain text formatters for artifacts. |
 | `_artifact_listing` | [`_artifact_listing.py`](../src/notebooklm/_artifact_listing.py) | Listing and filtering operations for notebook artifacts. |
-| `_row_adapters` | [`_row_adapters.py`](../src/notebooklm/_row_adapters.py) | Wire-shape adapters that wrap raw batchexecute rows (`ArtifactRow`, etc.) behind named accessors so downloads, polling, and listing don't open-code positional indices. Soft-degrade and strict-mode behavior is pinned in `tests/unit/test_row_adapters.py`. |
+| `_row_adapters*` | [`_row_adapters.py`](../src/notebooklm/_row_adapters.py), [`_row_adapters_artifacts.py`](../src/notebooklm/_row_adapters_artifacts.py), [`_row_adapters_notes.py`](../src/notebooklm/_row_adapters_notes.py), [`_row_adapters_sources.py`](../src/notebooklm/_row_adapters_sources.py) | Wire-shape adapters that wrap raw batchexecute rows (`ArtifactRow`, `NoteRow`, `SourceRow`) behind named accessors so downloads, polling, and listing don't open-code positional indices. `_row_adapters.py` remains the compatibility re-export shim. Soft-degrade and strict-mode behavior is pinned in `tests/unit/test_row_adapters.py`. |
 | `_research_task_parser` | [`_research_task_parser.py`](../src/notebooklm/_research_task_parser.py) | Parses deep-research task results from raw rows. Returns dict-shaped output today; a typed-model migration is not yet complete. |
 | `_mutating_operations` | [`_mutating_operations.py`](../src/notebooklm/_mutating_operations.py) | Audit inventory binding each `PROBE_THEN_CREATE` registry entry to a `RecoveryKind` (`EXECUTABLE` or `DISABLE_ONLY`) with a reason. Cross-checked by the registry-audit unit test so a new `PROBE_THEN_CREATE` policy cannot land without either a recovery wrapper or a documented disable-only justification. |
 | `_types/` | [`_types/`](../src/notebooklm/_types) | Private package holding the dataclass and `Protocol` implementations behind the public `types.py` / per-feature public schemas. Split per domain (`artifacts.py`, `chat.py`, `notebooks.py`, `notes.py`, `sharing.py`, `sources.py`, plus `common.py` for shared shapes like `ConnectionLimits`). |
