+++ src/notebooklm/_source_listing.py
@@ -7,7 +7,7 @@ import logging
 from collections.abc import Awaitable, Callable
 from typing import Any
 
-from ._row_adapters import SourceRow
+from ._row_adapters_sources import SourceRow
 from ._session_contracts import RpcCaller
 from .rpc import RPCError, RPCMethod
 from .types import Source
