+++ tests/unit/test_row_adapters.py
@@ -1,4 +1,4 @@
-"""Tests for ``notebooklm._row_adapters`` (``ArtifactRow`` + ``NoteRow`` + ``SourceRow``).
+"""Tests for row-adapter behavior and ``notebooklm._row_adapters`` compatibility.
 
 The adapters centralise position knowledge for the ``LIST_ARTIFACTS``,
 ``GET_NOTES_AND_MIND_MAPS``, and source row shapes so consumers
@@ -9,7 +9,8 @@ The adapters centralise position knowledge for the ``LIST_ARTIFACTS``,
 open-coding ``data[2]`` / ``data[4]`` / ``data[15]`` / ``row[1][1]`` /
 ``row[1][4]`` / ``data[0][0]`` / ``metadata[4]``. See
 ``docs/improvement.md`` §6.2 for the motivation and
-``src/notebooklm/_row_adapters.py`` for the position contracts.
+``src/notebooklm/_row_adapters_*.py`` for the position contracts and
+``src/notebooklm/_row_adapters.py`` for the compatibility shim.
 
 These tests cover three layers per adapter:
 
@@ -635,7 +636,7 @@ class TestNoteRowPositionContract:
     These pin tests fail loudly if anyone edits a position constant.
     When that happens, the failing diff IS the audit trail for the
     Google-side wire reshape. See
-    ``src/notebooklm/_row_adapters.py:NoteRow`` for the shape contract.
+    ``src/notebooklm/_row_adapters_notes.py:NoteRow`` for the shape contract.
     """
 
     def test_id_position_is_0(self) -> None:
